requirements.txt
mcp[cli]
httpx
python-dotenv
infisicalsdk
